requirements.txt
apache-airflow-providers-http